CR: No underscores in package names when renaming imports
