+++ loadtest/local_load_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	db_params "github.com/treeverse/lakefs/db/params"
+	dbparams "github.com/treeverse/lakefs/db/params"
 	"github.com/treeverse/lakefs/dedup"
 
 	"github.com/ory/dockertest/v3"
@@ -16,7 +16,7 @@ import (
 	"github.com/treeverse/lakefs/auth"
 	"github.com/treeverse/lakefs/auth/crypt"
 	authmodel "github.com/treeverse/lakefs/auth/model"
-	auth_params "github.com/treeverse/lakefs/auth/params"
+	authparams "github.com/treeverse/lakefs/auth/params"
 	"github.com/treeverse/lakefs/block"
 	"github.com/treeverse/lakefs/catalog"
 	"github.com/treeverse/lakefs/db"
@@ -58,10 +58,10 @@ func TestLocalLoad(t *testing.T) {
 	conn, _ := testutil.GetDB(t, databaseURI)
 	blockAdapter := testutil.NewBlockAdapterByEnv(&block.NoOpTranslator{})
 	cataloger := catalog.NewCataloger(conn)
-	authService := auth.NewDBAuthService(conn, crypt.NewSecretStore([]byte("some secret")), auth_params.ServiceCache{})
+	authService := auth.NewDBAuthService(conn, crypt.NewSecretStore([]byte("some secret")), authparams.ServiceCache{})
 	retentionService := retention.NewService(conn)
 	meta := auth.NewDBMetadataManager("dev", conn)
-	migrator := db.NewDatabaseMigrator(db_params.Database{DatabaseURI: databaseURI})
+	migrator := db.NewDatabaseMigrator(dbparams.Database{DatabaseURI: databaseURI})
 	dedupCleaner := dedup.NewCleaner(blockAdapter, cataloger.DedupReportChannel())
 	t.Cleanup(func() {
 		// order is important - close cataloger channel before dedup
